Improve routing API
