--- Sources/VaporWalletOrders/VaporWalletOrders.docc/Extensions/OrdersService.md
@@ -1,12 +1,0 @@
-# ``VaporWalletOrders/OrdersService``
-
-## Topics
-
-### Essentials
-
-- ``build(order:on:)``
-- ``register(migrations:)``
-
-### Push Notifications
-
-- ``sendPushNotifications(for:on:)``
