+++ Sources/VaporWalletPasses/PassesService.swift
@@ -38,8 +38,6 @@ public final class PassesService<PD: PassDataModel>: Sendable where Pass == PD.P
     ///
     /// - Parameters:
     ///   - app: The `Vapor.Application` to use in route handlers and APNs.
-    ///   - pushRoutesMiddleware: The `Middleware` to use for push notification routes. If `nil`, push routes will not be registered.
-    ///   - logger: The `Logger` to use.
     ///   - pemWWDRCertificate: Apple's WWDR.pem certificate in PEM format.
     ///   - pemCertificate: The PEM Certificate for signing passes.
     ///   - pemPrivateKey: The PEM Certificate's private key for signing passes.
@@ -47,8 +45,6 @@ public final class PassesService<PD: PassDataModel>: Sendable where Pass == PD.P
     ///   - openSSLPath: The location of the `openssl` command as a file path.
     public init(
         app: Application,
-        pushRoutesMiddleware: (any Middleware)? = nil,
-        logger: Logger? = nil,
         pemWWDRCertificate: String,
         pemCertificate: String,
         pemPrivateKey: String,
@@ -57,8 +53,6 @@ public final class PassesService<PD: PassDataModel>: Sendable where Pass == PD.P
     ) throws {
         self.service = try .init(
             app: app,
-            pushRoutesMiddleware: pushRoutesMiddleware,
-            logger: logger,
             pemWWDRCertificate: pemWWDRCertificate,
             pemCertificate: pemCertificate,
             pemPrivateKey: pemPrivateKey,
@@ -114,5 +108,11 @@ public final class PassesService<PD: PassDataModel>: Sendable where Pass == PD.P
     ///   - db: The `Database` to use.
     public func sendPushNotifications(for pass: PD, on db: any Database) async throws {
         try await service.sendPushNotifications(for: pass, on: db)
+    }
+}
+
+extension PassesService: RouteCollection {
+    public func boot(routes: any RoutesBuilder) throws {
+        try service.boot(routes: routes)
     }
 }
