--- Sources/VaporWalletPasses/VaporWalletPasses.docc/Extensions/PassesService.md
@@ -1,13 +1,0 @@
-# ``VaporWalletPasses/PassesService``
-
-## Topics
-
-### Essentials
-
-- ``build(pass:on:)``
-- ``build(passes:on:)``
-- ``register(migrations:withPersonalization:)``
-
-### Push Notifications
-
-- ``sendPushNotifications(for:on:)``
